+++ packages/icon-button/src/test/mwc-icon-button.test.ts
@@ -82,18 +82,17 @@ suite('mwc-icon-button', () => {
 
   const svgTemplate = document.createElement('template');
   svgTemplate.innerHTML = `
-  <svg slot="icon" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M0 0h24v24H0z" fill="none"/><path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm-2 15l-5-5 1.41-1.41L10 14.17l7.59-7.59L19 8l-9 9z"/></svg>
-  <svg slot="offIcon" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path fill="none" d="M0 0h24v24H0V0zm0 0h24v24H0V0z"/><path d="M16.59 7.58L10 14.17l-3.59-3.58L5 12l5 5 8-8zM12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8z"/></svg>`;
+  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M0 0h24v24H0z" fill="none"/><path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm-2 15l-5-5 1.41-1.41L10 14.17l7.59-7.59L19 8l-9 9z"/></svg>`;
 
-  test('nodes with `slot=icon` will serve as the on icon', async () => {
-    const iconQuery = svgTemplate.content.querySelector('svg[slot="icon"]')!;
+  test('default node will serve as the on icon', async () => {
+    const iconQuery = svgTemplate.content.querySelector('svg')!;
     assert.instanceOf(iconQuery, SVGElement);
 
     const icon = iconQuery.cloneNode(true);
     element.appendChild(icon);
     await element.updateComplete;
     const root = element.shadowRoot!;
-    const iconSlot = root.querySelector('slot[name="icon"]') as HTMLSlotElement;
+    const iconSlot = root.querySelector('slot') as HTMLSlotElement;
     assert.include(iconSlot.assignedNodes(), icon);
   });
 });